[frost] Move proptest into /tests
So we don't accidentally use internal APIs
